Fixes #72 - Make hyperband oracle obey objective direction
The existing implementation of hyperband did not take into account the objective direction when sorting candidates during selection. The older implementation inverted the scores when necessary (which was rather ugly) and that bit appears to have been lost in the refactor. Thus, when a max-sorted objective was used, HyperbandOracle._select_candidates would sort in the wrong direction.

HyperbandTuner takes the objective name, and uses the infer_metric_direction_from_name introduced in #76, then passes the resulting direction as a constructor argument to the Oracle.

Added a unit test showing that max, min, and default (==min) sorting all now work.