Argument for the 'fit' method verbosity.
kwargs are queried for the 'keras_verbose' parameter
setting verbosity for the 'fit' method of the instantiated
model tested by the tuner.

Default logging level set to 1. As specified in
Keras documentation of the 'fit' method.

The 'tuner_correctness_test.test_build_and_fit_model_in_tuner'
was adjusted to pass the 'keras_verbose' to the 'run_trial' method.

Signed-off-by: Jiri Podivin <jpodivin@gmail.com>